backend/requirements.txt
asgiref
django
django-allauth
dj-rest-auth
djangorestframework
djangorestframework_simplejwt
djoser
django-cors-headers
django-extensions
psycopg
psycopg-binary
redis
requests
python-dotenv
PyJWT
social-auth-app-django
social-auth-core
sqlparse
tzdata
